requirements.txt
streamlit==1.44.1
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
plotly==5.24.1
requests==2.32.3
python-dotenv==1.0.1
